chore: reorganize finance routes to make v2 the default and move v1 to legacy
