+++ src/app/(main)/dashboard/(legacy)/finance-v1/_components/kpis/net-worth.tsx
