+++ src/app/(main)/dashboard/finance/_components/balance-distribution-card.tsx
