+++ src/app/(main)/dashboard/finance/page.tsx
@@ -1,55 +1,84 @@
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
+import { format } from "date-fns";
+import { Download, RotateCw, Settings2 } from "lucide-react";
 
-import { CardOverview } from "./_components/card-overview";
-import { CashFlowOverview } from "./_components/cash-flow-overview";
-import { IncomeReliability } from "./_components/income-reliability";
-import { MonthlyCashFlow } from "./_components/kpis/monthly-cash-flow";
-import { NetWorth } from "./_components/kpis/net-worth";
-import { PrimaryAccount } from "./_components/kpis/primary-account";
-import { SavingsRate } from "./_components/kpis/savings-rate";
-import { SpendingBreakdown } from "./_components/spending-breakdown";
+import { Button } from "@/components/ui/button";
+import { Tabs, TabsList, TabsTrigger } from "@/components/ui/tabs";
+
+import { BalanceDistributionCard } from "./_components/balance-distribution-card";
+import { FinanceNotification } from "./_components/finance-notification";
+import { IncomeBreakdown } from "./_components/income-breakdown";
+import { OverviewKpis } from "./_components/overview-kpis";
+import { QuickActions } from "./_components/quick-actions";
+import { TransactionsOverviewCard } from "./_components/transactions-overview-card";
+import { UpcomingTransactionsTable } from "./_components/upcoming-transactions-table";
+import { Wallet } from "./_components/wallet";
 
 export default function Page() {
+  const formattedDate = format(new Date(), "EEEE, do MMMM yyyy");
+
   return (
-    <div>
-      <Tabs className="gap-4" defaultValue="overview">
-        <TabsList>
-          <TabsTrigger value="overview">Overview</TabsTrigger>
-          <TabsTrigger disabled value="activity">
-            Activity
-          </TabsTrigger>
-          <TabsTrigger disabled value="insights">
-            Insights
-          </TabsTrigger>
-          <TabsTrigger disabled value="utilities">
-            Utilities
-          </TabsTrigger>
-        </TabsList>
+    <div className="flex flex-col gap-4">
+      <div className="space-y-1">
+        <h1 className="text-3xl tracking-tight">Personal Finances</h1>
+        <p className="text-muted-foreground text-sm">{formattedDate}</p>
+      </div>
 
-        <TabsContent value="overview">
-          <div className="flex flex-col gap-4 **:data-[slot=card]:shadow-xs">
-            <div className="grid grid-cols-1 gap-4 *:data-[slot=card]:gap-2 *:data-[slot=card]:bg-linear-to-t *:data-[slot=card]:from-primary/5 *:data-[slot=card]:to-card sm:grid-cols-2 lg:grid-cols-4 xl:grid-cols-4">
-              <PrimaryAccount />
-              <NetWorth />
-              <MonthlyCashFlow />
-              <SavingsRate />
-            </div>
+      <div className="flex flex-col gap-3 lg:flex-row lg:items-center lg:justify-between">
+        <Tabs defaultValue="30-days">
+          <TabsList variant="line">
+            <TabsTrigger value="30-days">Dashboard</TabsTrigger>
+            <TabsTrigger value="12-months">Accounts</TabsTrigger>
+            <TabsTrigger value="custom">Transactions</TabsTrigger>
+          </TabsList>
+        </Tabs>
 
-            <div className="grid grid-cols-1 gap-4 lg:grid-cols-[minmax(0,2fr)_minmax(320px,1fr)]">
-              <div className="flex flex-col gap-4">
-                <CashFlowOverview />
-
-                <div className="grid grid-cols-1 gap-4 lg:grid-cols-2">
-                  <SpendingBreakdown />
-                  <IncomeReliability />
-                </div>
-              </div>
-
-              <CardOverview />
-            </div>
+        <div className="flex flex-wrap items-center gap-3">
+          <div className="flex items-center gap-1.5 text-muted-foreground text-xs">
+            <RotateCw className="size-4" />
+            <span>Updated 5 min ago</span>
           </div>
-        </TabsContent>
-      </Tabs>
+          <Button size="sm" variant="outline">
+            <Settings2 />
+            Settings
+          </Button>
+          <Button size="sm" variant="outline">
+            <Download data-icon="inline-start" />
+            Export
+          </Button>
+        </div>
+      </div>
+
+      <div className="grid grid-cols-1 gap-4 xl:grid-cols-12">
+        <div className="xl:col-span-6">
+          <OverviewKpis />
+        </div>
+
+        <div className="flex flex-col gap-4 xl:col-span-6">
+          <IncomeBreakdown />
+          <FinanceNotification />
+        </div>
+      </div>
+
+      <div className="grid grid-cols-1 gap-4 xl:grid-cols-12">
+        <div className="xl:col-span-7">
+          <TransactionsOverviewCard />
+        </div>
+        <div className="xl:col-span-5">
+          <BalanceDistributionCard />
+        </div>
+      </div>
+
+      <div className="grid grid-cols-1 gap-4 xl:grid-cols-12">
+        <div className="xl:col-span-4">
+          <Wallet />
+        </div>
+        <div className="xl:col-span-4">
+          <UpcomingTransactionsTable />
+        </div>
+        <div className="xl:col-span-4">
+          <QuickActions />
+        </div>
+      </div>
     </div>
   );
 }
